requirements.txt
Flask>=3.0.0
PyJWT>=2.8.0
marshmallow>=3.20
python-dotenv>=1.0
